refactor: make transaction details more simple as it was doing too much
